% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RZonar-package.R
\docType{package}
\name{RZonar-package}
\alias{RZonar}
\alias{RZonar-package}
\title{RZonar: Conveniently access the Zonar API from R}
\description{
The Zonar API documented at https://support.zonarsystems.net/hc/en-us enables the programatic retrieval of Zonar data. This R package wraps that API, making it more convenient to retrieve both specific and bulk data from the Zonar API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BostonPublicSchools/RZonar}
  \item \url{https://bostonpublicschools.github.io/RZonar/}
  \item Report bugs at \url{https://github.com/BostonPublicSchools/RZonar/issues}
}

}
\keyword{internal}
