% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{zonar_get_assets}
\alias{zonar_get_assets}
\alias{zonar_get_gps_units}
\alias{zonar_get_asset_usage}
\title{Get Zonar asset (bus) data}
\usage{
zonar_get_assets(assets = NULL)

zonar_get_gps_units()

zonar_get_asset_usage()
}
\arguments{
\item{assets}{Character vector of asset ids to retrieve. If \code{NULL} (the default)
retrieve all assets. If not \code{NULL}, a character vector of the must be the
Zonar \emph{Asset IDs} to retrieve.}
}
\description{
\code{zonar_get_assets} retrieves zonar asset (bus) information from
the Zonar API
}
\section{Functions}{
\itemize{
\item \code{zonar_get_gps_units()}: Get gps info

\item \code{zonar_get_asset_usage()}: Get asset usage

}}
\examples{
\dontshow{if (Sys.getenv("ZONAR_CUSTOMER")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

assettestall <- zonar_get_assets()
dplyr::glimpse(assettestall)

assettest <- zonar_get_assets(assets = c("586", "1046"))
dplyr::glimpse(assettest)
\dontshow{\}) # examplesIf}
}
