% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{zonar_get_paths}
\alias{zonar_get_paths}
\title{Get zonar schedule data for a specific zone}
\usage{
zonar_get_paths(start, end, assets = NULL, timezone = "America/New_York")
}
\arguments{
\item{start}{Character vector of length one giving the start time in
yyyy-mm-dd hh:mm:ss format.}

\item{end}{Character vector of length one giving the end time in
yyyy-mm-dd hh:mm:ss format.}

\item{assets}{Character vector of Zonar asset ids to retrieve. If \code{NULL}
(the default) retrieve paths for all assets.}

\item{timezone}{Defaults to "America/New_York"}
}
\description{
Get zonar schedule data for a specific zone
}
\examples{
\dontshow{if (Sys.getenv("ZONAR_CUSTOMER")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(RZonar)

pathtest_all <- zonar_get_paths(
  start = "2022-11-30 08:10:00",
  end = "2022-11-30 08:40:00")
dplyr::glimpse(pathtest_all)
pathtest_586 <- zonar_get_paths(
  start = "2022-11-30 08:10:00",
  end = "2022-11-30 08:40:00",
  assets = "586")
dplyr::glimpse(pathtest_586)
\dontshow{\}) # examplesIf}
}
