% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{zonar_get_schedules}
\alias{zonar_get_schedules}
\title{Get zonar schedule data for all routes}
\usage{
zonar_get_schedules(
  start,
  end,
  by = c("time", "zone", "asset"),
  zones = NULL,
  assets = NULL,
  timezone = "America/New_York",
  reformat = FALSE,
  include_geometry = FALSE,
  cachedir = NULL,
  omit_categories = NULL,
  test = FALSE
)
}
\arguments{
\item{start}{Character vector of length one giving the start time in
yyyy-mm-dd hh:mm:ss format.}

\item{end}{Character vector of length one giving the end time in
yyyy-mm-dd hh:mm:ss format.}

\item{by}{Whether to split requests by time or zone under the hood. By default
a reasonable value is chosen based on the amount of data you request (i.e., the
size of the interval between \code{start} and \code{end}.}

\item{zones}{Character vector of zone \strong{names} (not IDs!) to retrieve. If \code{NULL} (the default)
retrieve all zones.}

\item{assets}{Integer vector of giving the Zonar asset \strong{IDs} (not names) of the asset
to retrieve.}

\item{timezone}{Defaults to "America/New_York"}

\item{reformat}{If \code{TRUE} cleanup and re-format the data such that each row
represents a "Zone event", i.e., a discrete entry into and/or exit from a
zone. Otherwise leave the date in the original format returned by Zonar.
Defaults to \code{FALSE}.}

\item{include_geometry}{If \code{TRUE} also return the geographic polygon
showing the location and extent of each Zone. Defaults to \code{FALSE}.}

\item{cachedir}{Ignored, always uses .zonarCache now.}

\item{omit_categories}{Character vector containing zone categories to skip/omit.}

\item{test}{Set to \code{TRUE} for testing purposes, otherwise it will be too slow.}
}
\description{
Use \code{zonar_get_schedules} to retrieve the time each bus
(asset) reaches a route anchor (zone) from the
\href{https://support.zonarsystems.net/hc/en-us/categories/360000020371-API-Portal}{Zonar API}
}
\details{
The Zonar
\href{https://support.zonarsystems.net/hc/en-us/articles/360018372852-GPS-Schedule-Report}{schedule report}
API can give us that information, but not all at once. Using this R wrapper
makes it easy to retrieve bulk schedule reports from the Zonar API in a
convenient format that can easily be saved to a database or spreadsheet.

Note that retrieving data for short intervals (one day or less) will generally
be faster when setting \code{by = "time"}, while bulk retrieval of weeks or months
of data will be faster when setting \code{by = "zone"}. By default \code{zonar_get_shedules_all}
will try to pick the best setting for you depending on the \code{start} and \code{end}
parameters you give it.

The default is to retrieve all zones and all assets for the time period specified
by \code{start} and \code{end}. You can retrieve a specified list of zones or assets
via the \code{zones} and \code{assets} arguments. However, due to the way the Zonar API
works you cannot specify both. Note that \code{zones} must be specified by name, but
\code{assets} must be specified by ID; see \code{\link[=zonar_get_assets]{zonar_get_assets()}} if you need to look
up IDs from asset names.
}
\section{Functions}{
\itemize{
\item \code{zonar_get_schedules()}: Get zonar schedule data for all routes

}}
\examples{
\dontshow{if (Sys.getenv("ZONAR_CUSTOMER")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(RZonar)
## use by="time" for short duration, e.g. < 1 day
schedtest1 <- zonar_get_schedules(
  start = "2022-11-30 08:00:00",
  end = "2022-11-30 08:40:00",
  by = "time")
dplyr::glimpse(schedtest1)

## use by="zone" for bulk retrieval (e.g. weeks or months of data)
schedtest2 <- zonar_get_schedules(
  start = "2022-11-20 08:00:00",
  end = "2022-11-28 08:30:00",
  by = "zone",
  zones = c("Adams", "Burke") # not required, leave out for all zones
  )
dplyr::glimpse(schedtest2)

## use by="zone" for bulk retrieval (e.g. weeks or months of data)
schedtest3 <- zonar_get_schedules(
  start = "2022-11-20 08:00:00",
  end = "2022-11-28 08:30:00",
  assets = c("1024", "1030") # not required, leave out for all assets
  )
dplyr::glimpse(schedtest3)
\dontshow{\}) # examplesIf}
}
