% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zonar_get_zones}
\alias{zonar_get_zones}
\title{Get Zonar Zone / Location-of-Interest data}
\usage{
zonar_get_zones(zone = NULL)
}
\arguments{
\item{zone}{Optional zone name to retrieve. If \code{NULL} (the default) retrieve all zones.}
}
\description{
\code{zonar_get_zones} retrieves zonar "locations of interest" AKA zones
}
