% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpass.R
\name{zonar_get_zpass}
\alias{zonar_get_zpass}
\title{Get zonar zpass tapcard data}
\usage{
zonar_get_zpass(start, end, timezone = "America/New_York", convert = TRUE)
}
\arguments{
\item{start}{Character vector of length one giving the start time in
yyyy-mm-dd hh:mm:ss format.}

\item{end}{Character vector of length one giving the end time in
yyyy-mm-dd hh:mm:ss format.}

\item{timezone}{Defaults to "America/New_York"}

\item{convert}{If \code{TRUE} convert data types, otherwise keep returned values as character}
}
\value{
A \code{data.frame} containing all available tap card data.
}
\description{
Retrieves tap-card data for the specified time period from the Zonar API
}
