% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zonar_rename_zones}
\alias{zonar_rename_zones}
\title{Get Zonar Zone / Location-of-Interest data}
\usage{
zonar_rename_zones(old, new)
}
\arguments{
\item{old}{Character vector of existing zone names}

\item{new}{Character vector of replacement names}
}
\description{
\code{zonar_rename_zones} renames zonar "locations of interest" AKA zones
}
