% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zonar_req}
\alias{zonar_req}
\title{Create base Zonar API request}
\usage{
zonar_req(...)
}
\arguments{
\item{...}{credential overrides, passed to \code{zonar_credentials}. Usually leave
this empty and set login info as environment variables.}
}
\description{
Create basic request, can be reused.
}
